requirements.txt
Django==1.10.1
djangorestframework==3.5.3

// django-filters==0.2.1
django-filter==0.15.3
djangorestframework-filters==0.9.1
